requirements.txt
Flask==2.0.2
numpy==1.21.2
Pillow==8.3.2
requests==2.26.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cpu
torchvision==0.10.1+cpu